+++ templates/feeback.html
@@ -1,0 +1,104 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Feedback</title>
+  <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" rel="stylesheet">
+  <style>
+    body {
+      margin: 0;
+      font-family: Arial, sans-serif;
+      background-color: #f0f0f0;
+      padding: 50px;
+    }
+
+    .feedback-container {
+      background-color: white;
+      padding: 30px;
+      border-radius: 10px;
+      box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+      max-width: 600px;
+      margin: auto;
+    }
+
+    .feedback-container h2 {
+      margin-bottom: 20px;
+    }
+
+    .feedback-container input,
+    .feedback-container textarea {
+      width: 100%;
+      padding: 10px;
+      margin: 10px 0;
+      border: 1px solid #ccc;
+      border-radius: 5px;
+    }
+
+    .feedback-container button {
+      font-size: 1em;
+      padding: 10px 20px;
+      background-color: #6a11cb;
+      color: white;
+      border: none;
+      border-radius: 5px;
+      transition: background-color 0.3s ease, transform 0.3s ease;
+    }
+
+    .feedback-container button:hover {
+      background-color: #2575fc;
+      transform: scale(1.05);
+    }
+  </style>
+</head>
+<body>
+  <div class="feedback-container">
+    <h2>Feedback</h2>
+    <form id="feedbackForm">
+      <div class="form-group">
+        <input type="email" id="userEmail" placeholder="Your email address" required>
+      </div>
+      <div class="form-group">
+        <textarea id="userFeedback" rows="5" placeholder="Your feedback" required></textarea>
+      </div>
+      <button type="button" onclick="sendFeedback()">Submit Feedback</button>
+    </form>
+  </div>
+
+  <!-- Email.js script -->
+  <script src="https://cdn.jsdelivr.net/npm/@emailjs/browser@3.11.0/dist/email.min.js"></script>
+  <script>
+    (function() {
+      emailjs.init("jx8TiUIaNuP77fD7N"); // Ensure this public key is correct
+    })();
+
+    function sendFeedback() {
+      const userEmail = document.getElementById("userEmail").value;
+      const userFeedback = document.getElementById("userFeedback").value;
+      
+      if (validateEmail(userEmail) && userFeedback.trim() !== "") {
+        emailjs.send("service_hlql7em", "template_n7qzpvm", {
+          userEmail: userEmail,
+          userFeedback: userFeedback
+        })
+        .then((response) => {
+          console.log("SUCCESS!", response.status, response.text);
+          alert("Feedback sent successfully!");
+          document.getElementById("feedbackForm").reset();
+        })
+        .catch((error) => {
+          console.error("Error:", error); 
+          alert("Failed to send feedback. Please try again later.");
+        });
+      } else {
+        alert("Please enter a valid email address.");
+      }
+    }
+
+    function validateEmail(email) {
+      const re = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
+      return re.test(String(email).toLowerCase());
+    }
+  </script>
+</body>
+</html>
